Minor fixes

+++ update.ps1
@@ -2,9 +2,15 @@ $current = Split-Path -Parent $PSCommandPath
 
 Write-Host "Updating application..."
 
-Remove-Item -Recurse -Force "$current\assets" | Out-Null
-Remove-Item -Recurse -Force "$current\index.html" | Out-Null
-Remove-Item -Recurse -Force "$current\update.ps1" | Out-Null
+If (Test-Path "$current\assets") {
+    Remove-Item -Recurse -Force "$current\assets" | Out-Null
+}
+If (Test-Path "$current\index.html") {
+    Remove-Item -Recurse -Force "$current\index.html" | Out-Null
+}
+If (Test-Path "$current\update.ps1") {
+    Remove-Item -Recurse -Force "$current\update.ps1" | Out-Null
+}
 
 Invoke-WebRequest -Uri "https://github.com/braathen/qlik-hardware-configurator/archive/master.zip" -UseBasicParsing -OutFile "./master.zip"
